jets: remove ut_fish
This jet was previously decapitated in 61b957cc5b1, meaning that its
logic was stubbed out and it simply passes through a cache.

My understanding is that only compiler jets that involve worklist
algorithms need to be cached in this way, and yet it doesn't appear to
me that +fish uses any such algorithms (it instead appears to maintain
its own internal cache).  Thus, I've removed the caching jet.